Move listeners to a handler class
